fix(mcp): fetched prompts lose image content
