+++ src/agents/agent-bundle-mcp-tools.materialize.test.ts
@@ -633,112 +633,6 @@ describe("createBundleMcpToolRuntime", () => {
     expect(runtime.diagnostics).toEqual(diagnostics);
   });
 
-  it("exposes MCP resource and prompt utility tools when advertised", async () => {
-    const base = makeToolRuntime({ tools: [], serverName: "knowledge" });
-    const publicResults = {
-      prompts_get: {
-        description: "Brief the user",
-        messages: [
-          {
-            role: "user",
-            content: {
-              type: "text",
-              text: "Summarize MCP",
-              annotations: { audience: ["assistant"] },
-              _meta: { promptBlock: "preserved" },
-            },
-          },
-        ],
-      },
-      prompts_list: {
-        prompts: [{ name: "brief", _meta: { promptEntry: "preserved" } }],
-        nextCursor: "prompt-page-two",
-      },
-      resources_list: {
-        resources: [
-          {
-            uri: "memo://one",
-            name: "memo",
-            annotations: { priority: 0.5 },
-            _meta: { resourceEntry: "preserved" },
-          },
-        ],
-        nextCursor: "resource-page-two",
-      },
-      resources_read: {
-        contents: [{ uri: "memo://one", text: "memo text", _meta: { content: "preserved" } }],
-      },
-    };
-    const privateResults = Object.fromEntries(
-      Object.entries(publicResults).map(([operation, value]) => [
-        operation,
-        { ...value, _meta: { privateState: `${operation}-must-not-leak` } },
-      ]),
-    );
-    const runtime = await materializeBundleMcpToolsForRun({
-      runtime: {
-        ...base,
-        getCatalog: async () => ({
-          version: 1,
-          generatedAt: 0,
-          servers: {
-            knowledge: {
-              serverName: "knowledge",
-              safeServerName: "knowledge",
-              launchSummary: "knowledge",
-              toolCount: 0,
-              resources: { listChanged: true },
-              prompts: { listChanged: true },
-            },
-          },
-          tools: [],
-        }),
-        listResources: async () => privateResults.resources_list,
-        readResource: async () => privateResults.resources_read,
-        listPrompts: async () => privateResults.prompts_list,
-        getPrompt: async () => privateResults.prompts_get,
-      },
-    });
-
-    expect(runtime.tools.map((tool) => tool.name)).toEqual([
-      "knowledge__prompts_get",
-      "knowledge__prompts_list",
-      "knowledge__resources_list",
-      "knowledge__resources_read",
-    ]);
-
-    for (const [operation, args] of [
-      ["prompts_get", { name: "brief" }],
-      ["prompts_list", {}],
-      ["resources_list", {}],
-      ["resources_read", { uri: "memo://one" }],
-    ] as const) {
-      const tool = expectDefined(
-        runtime.tools.find((candidate) => candidate.name === `knowledge__${operation}`),
-        `${operation} utility tool`,
-      );
-      const result = await tool.execute(`call-${operation}`, args, undefined, undefined);
-      expectTextContentBlock(result.content[0], JSON.stringify(publicResults[operation], null, 2));
-      expect(result.details).toMatchObject({
-        mcpServer: "knowledge",
-        mcpOperation: operation,
-        untrustedMcpOutput: true,
-      });
-      expect(tool.resultContentSource).toBe("network");
-      expect(expectDefined(privateResults[operation], `${operation} private source`)._meta).toEqual(
-        {
-          privateState: `${operation}-must-not-leak`,
-        },
-      );
-    }
-
-    await expect(
-      runtime.tools
-        .find((tool) => tool.name === "knowledge__prompts_get")!
-        .execute("call-prompt", { name: "brief", arguments: { count: 1 } }, undefined, undefined),
-    ).rejects.toThrow("arguments.count must be a string");
-  });
-
   it("applies per-server MCP tool filters to resource and prompt utility tools", async () => {
     const base = makeToolRuntime({ tools: [], serverName: "knowledge" });
     const runtime = await materializeBundleMcpToolsForRun({
